mobile: GlobalBottomSheet, dynamic sizing
